devices: Rename "htm_file" module to "file_parser"
We also want to parse other data files residing on the device. Change
the name of this module to make it more general as there are other files
beyond the "htm_file" we're interested in.
