+++ src/mbed_tools/devices/_internal/file_parser.py
